+++ .gitHooks/setupGitHooks.mjs
@@ -1,4 +1,4 @@
-import { process } from 'node:process';
+import process from 'node:process';
 import { execSync } from 'node:child_process';
 import { existsSync } from 'node:fs';